requirements.txt
asgiref==3.4.1
cycler==0.10.0
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
psycopg2==2.9.1
pyparsing==3.0.1
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
